RavenDB-5006: Add page locator usage in a few places and set a reasonable page locator size for VariableSizeTree.
In the future, we may benefit from allowing a page locator size of 32 -or higher- in read transactions if we manage to optimize PageLocator. See http://issues.hibernatingrhinos.com/issue/RavenDB-5320

+++ src/Voron/Data/RawData/ActiveRawDataSmallSection.cs
@@ -208,7 +208,7 @@ namespace Voron.Data.RawData
             sectionHeader->NumberOfEntries = 0;
             sectionHeader->NumberOfPages = numberOfPagesInSmallSection;
             sectionHeader->LastUsedPage = 0;
-            sectionHeader->SectionOwnerHash = Hashing.XXHash64.Calculate(owner.Content.Ptr, owner.Content.Length);
+            sectionHeader->SectionOwnerHash = Hashing.XXHash64.Calculate(owner.Content.Ptr, (ulong)owner.Content.Length);
 
             var availablespace = (ushort*)((byte*)sectionHeader + ReservedHeaderSpace);
